requirements.txt
requests>=2.31.0
pandas>=2.0.0
PyYAML>=6.0
gspread>=5.7.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
selenium>=4.10.0
slack-sdk>=3.21.2
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.9
